Entity System & Entity Components
This base class describes every player interactable object in the game and contains various properties about that object. For example of the entity is on fire or is damaged or has exploded. For example a vehicles, characters, destructible objects.

Every entity is now managed by a central class that handled the deletion and creation of the entities as well as damage and keeping track of various gameplay functions.

-Most objects are now a child class of Entity.cs